remove last changes
